KAFKA-4247; Prevent CLASSPATH from beginning with a single colon
Different fix for problem addressed by https://github.com/apache/kafka/pull/1953. Should prevent the CLASSPATH environment variable from being prefixed by a single colon before the JVM is invoked in the run-class script, which will then prevent the current working directory from being unintentionally included in the classpath when using the Reflections library.

If the current working directory should still be included in the classpath, it just needs to be explicitly specified either with its fully-qualified pathname or as a single dot (".").

Author: Chris Egerton <chrise@confluent.io>

Reviewers: Randall Hauch <rhauch@gmail.com>, Jun Rao <junrao@gmail.com>

Closes #4406 from C0urante/fix-colon-prefixed-classpath