MINOR: Replace empty string concattenation with String.valueOf() where possible. (#4369)
This is something I did after my working hours, I would ask people reviewing this do the same, don't take time for this during your work hours.

I try to keep such a PR as limited as possible, for clarity of reading.

==========

Using an empty string concat in order to achieve the String representation of the value you want is bad for 2 reasons, as explained here: (https://stackoverflow.com/questions/1572708/is-conversion-to-string-using-int-value-bad-practice

Readability: it shows what you're trying to do.

Depending on your compiler, it might attempt to create your String by first creating a StringBuffer, appending your value to it and then doing .toString() on that. Which is inefficient.

Also, the Metrics.java file had an empty string being added for the sole reason that the page width forced a string to continue on a new line. Removed that.

Reviewers: Guozhang Wang <wangguoz@gmail.com>