KAFKA-6492; Fix log truncation to empty segment
This patch ensures that truncation to an empty segment forces resizing of the index file in order to prevent premature rolling.

I have added unit tests which verify that appends are permitted following truncation to an empty segment. Without the fix, this test case reproduces the failure in which the rolled segment matches the current active segment.

Author: Jason Gustafson <jason@confluent.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>, Jiangjie (Becket) Qin <becket.qin@gmail.com>

Closes #4498 from hachikuji/KAFKA-6492