HOTFIX: Fix lgtm.com alerts (dead code and out-of-bounds error) (#4388)
This fixes two alerts flagged on lgtm.com for Apache Kafka.

This dead code alert where InvalidTypeIdException indirectly extends JsonMappingException. The flagged condition with the type test appears after the type test for the latter and thus makes its body dead. I opted to change the order of the tests. Please let me know if this is the intended behavior.

The second commit addresses this out-of-bounds alert.

More alerts can be found here. Note that my colleague Aditya Sharad addressed some of those in the now outdated #2939.

Reviewers: Matthias J. Sax <matthias@confluent.io>, Rajini Sivaram <rajinisivaram@googlemail.com>