MINOR: Free sends in MultiSend as they complete (#4574)
Currently we hold onto all Records references in a multi-partition fetch response until the full response has completed. This can be a problem when the records have been down-converted since they will be occupying a (potentially large) chunk of memory. This patch changes the behavior in MultiSend so that once a Send is completed, we no longer keep a reference to it, which will allow the Records objects to be freed sooner.

I have added a simple unit test to verify that sends are removed as the MultiSend progresses.

Reviewers: Ismael Juma <ismael@juma.me.uk>