KAFKA-6773; Allow offset commit/fetch/describe/delete with empty groupId (#4851)
We had a regression in #4788 which caused the offset commit/fetch/describe APIs to fail if the groupId was empty. This should be allowed for backwards compatibility. Additionally, I have modified DeleteGroups to allow removal of the empty group, which was missed in the initial implementation. I've added a test case to ensure that we do not miss this again in the future.

Reviewers: Ismael Juma <ismael@juma.me.uk>